move duplicated code to the base protocol
