+++ sphire/objects.py
@@ -24,8 +24,6 @@
 # *
 # **************************************************************************
 
-import os
-
 import pyworkflow.object as pwobj
 from pwem import EMObject
